fix: 게시글 목록 조회, 상세 조회 페이지에 댓글수 오류 해결

+++ server/src/main/java/io/github/ysbunny/community/post/dto/response/PostListItemResponse.java
@@ -14,7 +14,7 @@ public class PostListItemResponse {
     private String authorNickname;
     private String authorProfileImage;
 
-    private Integer commentCount;
+    private Long commentCount;
 
     private LocalDateTime createdAt;
 }
